refactor: extract input props to separate file
