+++ src/Shared/Components/Input/index.ts
@@ -1,1 +1,2 @@
 export * from "./Input";
+export * from "./propsInput";
