+++ src/Shared/Components/Input/propsInput.ts
@@ -1,0 +1,8 @@
+export interface InputProps
+  extends Omit<React.InputHTMLAttributes<HTMLInputElement>, "onChange"> {
+  value: string;
+  disabled?: boolean | undefined;
+  onChange: (value: string) => void;
+  placeholder?: string;
+  className?: string;
+}
